requirements.txt
openai==1.43.0
python-dotenv==1.0.1
tiktoken==0.7.0
